Revert "randao mixes"
This reverts commit 2863f9c24dd8cfc44ffce16321f68ef3b793e9a2.

+++ beacon-chain/core/epoch/epoch_processing.go
@@ -289,7 +289,7 @@ func ProcessFinalUpdates(state *stateTrie.BeaconState) (*stateTrie.BeaconState, 
 
 	// Set RANDAO mix.
 	randaoMixLength := params.BeaconConfig().EpochsPerHistoricalVector
-	if state.RandaoMixesLength() != randaoMixLength {
+	if uint64(state.RandaoMixesLength()) != randaoMixLength {
 		return nil, fmt.Errorf(
 			"state randao length %d different than EpochsPerHistoricalVector %d",
 			state.RandaoMixesLength(),
